requirements.txt
# Streamlit for UI
streamlit

# Data handling
pandas
numpy

# Environment variables
python-dotenv

# Langchain and related packages
langchain
langchain_experimental
langchain-google-genai
langchain-core
faiss-cpu

# Additional dependencies for error handling
tenacity
